lab5 praca w domu ;)

+++ src/pl/edu/pw/elka/prm2t/lab4/zad2/main.java
@@ -44,7 +44,7 @@ public class main {
                     continue;
                 }
                 studenci.put(nazwisko + " " + nrindeksu, tab);
-                srednie.put(nazwisko, srednia);
+                srednie.put(nazwisko, (int) srednia);
             }
             else{
                 System.out.println("koniec");
